# Conflicts:
#	Valkyrien-Ship-Schematics
#	common/build.gradle
#	common/src/main/java/net/spaceeye/vmod/mixin/SplitShipMixin.java
#	common/src/main/kotlin/net/spaceeye/vmod/gui/SimpleMessager.kt
#	common/src/main/kotlin/net/spaceeye/vmod/translate/Translatable.kt
#	fabric/build.gradle
#	forge/build.gradle
#	forge/src/main/java/net/spaceeye/vmod/forge/mixin/TakeoffBearingMixin.java
#	forge/src/main/kotlin/net/spaceeye/vmod/forge/TakeoffMixinCompat.kt
#	gradle.properties

+++ common/src/main/kotlin/net/spaceeye/vmod/constraintsManaging/extensions/SignalActivator.kt
@@ -25,9 +25,9 @@ class SignalActivator(): MConstraintExtension, TickableMConstraintExtension {
         this.percentageNameReflection = percentageNameReflection
     }
 
-    var targetPercentage = 0.0
+    var targetPercentage = 0.0f
 
-    var totalPercentage = 0.0
+    var totalPercentage = 0.0f
     var numMessages = 0
 
     private var wasDeleted = false
@@ -37,7 +37,7 @@ class SignalActivator(): MConstraintExtension, TickableMConstraintExtension {
     private fun signalTick(msg: Message) {
         if (msg !is Signal) { return }
 
-        totalPercentage = min(max(msg.percentage, 0.0), 1.0)
+        totalPercentage = min(max(msg.percentage.toFloat(), 0.0f), 1.0f)
         numMessages++
     }
 
@@ -45,7 +45,7 @@ class SignalActivator(): MConstraintExtension, TickableMConstraintExtension {
         if (numMessages != 0) {
             targetPercentage = totalPercentage / numMessages
             numMessages = 0
-            totalPercentage = 0.0
+            totalPercentage = 0.0f
         }
 
         percentageField.set(obj, targetPercentage)
